feat: create useAuth and refactor frontend
