+++ frontend/package.json
@@ -16,6 +16,7 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
+    "sass": "^1.54.9",
     "socket.io-client": "^4.5.1",
     "web-vitals": "^2.1.4"
   },
