--- frontend/src/components/Navbar/Navbar.css
@@ -1,7 +1,0 @@
-.navbar_design {
-    font-family: 'monospace';
-    font-weight: 700;
-    letter-spacing: '.3rem';
-    color: 'inherit';
-    text-decoration: 'none';
-}
