+++ frontend/src/containers/App/App.module.scss
@@ -1,0 +1,10 @@
+.screen {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  height: 100%;
+}
+
+.main {
+  height: 100%;
+}
