+++ frontend/src/containers/Dashboard/index.js
@@ -1,0 +1,13 @@
+import { Box, Typography } from '@mui/material';
+
+const Dashboard = () => {
+  return (
+    <div>
+      <Box display={'flex'} flexDirection={'column'} width={'100%'} alignItems="center">
+        <Typography>I am the dashboard</Typography>
+      </Box>
+    </div>
+  );
+};
+
+export default Dashboard;
