+++ frontend/src/containers/LevelSelect/index.js
@@ -1,0 +1,27 @@
+import { Box, FormControl, MenuItem, Select, Typography } from '@mui/material';
+
+const LevelSelect = () => {
+  const handleChange = () => {
+    console.log('pressed');
+  };
+
+  return (
+    <Box display={'flex'} flexDirection={'column'} width={'100%'} alignItems={'center'}>
+      <FormControl fullWidth>
+        <Typography>Choose your preferred difficulty level:</Typography>
+        <Select
+          labelId="difficulty-level"
+          id="difficulty-select"
+          label="Age"
+          onChange={handleChange}
+        >
+          <MenuItem value="Easy">Easy</MenuItem>
+          <MenuItem value="Medium">Medium</MenuItem>
+          <MenuItem value="Hard">Hard</MenuItem>
+        </Select>
+      </FormControl>
+    </Box>
+  );
+};
+
+export default LevelSelect;
